+++ modules/grading_client/lib/grading_client/answers.ex
@@ -21,11 +21,13 @@ defmodule GradingClient.Answers do
 
     {answers, _} = Code.eval_file(filename)
 
-    Enum.each(answers, fn answer ->
-      :ets.insert(@table, {{answer.module_id, answer.question_id}, answer})
-    end)
+    modules =
+      MapSet.new(answers, fn answer ->
+        :ets.insert(@table, {{answer.module_id, answer.question_id}, answer})
+        answer.module_id
+      end)
 
-    {:ok, nil}
+    {:ok, %{modules: modules}}
   end
 
   @doc """
@@ -34,6 +36,19 @@ defmodule GradingClient.Answers do
   @spec check(integer(), integer(), String.t()) :: :correct | {:incorrect, String.t()}
   def check(module_id, question_id, answer) do
     GenServer.call(__MODULE__, {:check, module_id, question_id, answer})
+  end
+
+  @doc """
+  Returns the list of modules.
+  """
+  @spec get_modules() :: [atom()]
+  def get_modules() do
+    GenServer.call(__MODULE__, :get_modules)
+  end
+
+  @impl true
+  def handle_call(:get_modules, _from, state) do
+    {:reply, state.modules, state}
   end
 
   @impl true